[Fix] Add Check for Valid Target to ABC Buffer
If a player spams an action after their target dies, the targetBuffer would be updated to null causing the ABC Alert to trigger even though they didn't have a valid target.  This adds a sanity check to make sure that the targetid is valid before it is pushed to the buffer.