[Enh] Add Always-Be-Casting Alert
Adds an Always-Be-Casting alert that warns the player if they allow the GCD to expire without starting a new cast.  Includes some attempts at checking to make sure the target hasn't died or changed, in which case we expect/don't care that there is a gap in the GCD.  Also reorders the borders for the GCDBar to keep them always on top.