[Fix] Clip/ABC Color Reset/Timeout
If ShowOutOfCombat or ShowOnlyGCDRunning were enabled, DrawGCDWheel and DrawGCDBar were never invoked when the bar wasn't displayed.  As such, the code that reset the bar to the default background color wasn't evaluated.  This fix reduces the time for the reset to fire from a fixed 4s to 50ms less than whatever the current GCDTimeout is set to, ensuring that the bar/wheel is always reset to default just before disappearing.

Has the side effect of changing the default behavior when ShowOutOfCombat(ON) and ShowOnlyGCDRunning(OFF) by lowering the default timeout from 4s to 2s, but this is now adjustable with a GCD Timeout slider in the config for bar/wheel respectively.