[Enh] Implement IsRunning -> Refactor
This commit adds a "proper" IsRunning bool and uses it to implement a pseudo-timer based on iterations through Update().  ShowABCAlert and ShowOnlyGCDRunning are refactored to use this timer instead of the previous methods, avoiding any calls to the System Time.  SecondsSinceGCDEnd is removed as it is rendered obsolete by having a proper IsRunning bool.

The pseudo-timer is likely PC-performance dependent (e.g. will tick faster/slower based on framerate).  ABC alert sensitivity slider added to allow the player to tune to taste based on their PC and class (classes with cast-times larger than the recast time (e.g. Black Mage) will want less sensitivity).

To-do: investigate if it is worthwhile to sync with the System Time to have the ability to set a defined time.  Likely not worth it.