[Fix] Re-Add Castbar Contents RegEx
This is actually required to remove some junk the game adds to the string when using items (for example, learning a mount).  I presume the game uses it  for formatting (italics/bold/whatever) but in any case we don't want a bunch of junk text on our castbar so we have to filter it out.